requirements.txt
Flask==3.0.0
pyjwt==2.3.0
cryptography==3.4.8
pandas==1.3.5
numpy==1.26.3
requests==2.25.1
urllib3==1.26.5